Extract AppLifecycleCoordinator for app observer lifecycle
Move the ProcessLifecycleOwner foreground/background dispatch out of App
into a dedicated AppLifecycleCoordinator driving the stream and Hyperliquid
observers, so adding observers no longer touches App. Drop the unused
SyncPerpetualPositions injection from App (no startup side effect; it is
already injected where it is used).
